% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.bancadria-package.R
\docType{package}
\name{bankr.bancadria-package}
\alias{bankr.bancadria}
\alias{bankr.bancadria-package}
\title{bankr.bancadria: Reads and Tidy Bank Statements from Banca Adria}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea.ranzato01@gmail.com}

}
\keyword{internal}
